Refactor python examples module into a package
