+++ python/graphframes/examples/__init__.py
@@ -1,0 +1,5 @@
+
+from .belief_propagation import BeliefPropagation
+from .graphs import Graphs
+
+__all__ = ['BeliefPropagation', 'Graphs']
